Added commit on the varity of resolutions for the human component (EHC)
The requisite files and namelist defaults have been added for
the EHC at r025 (quarter degree) resolution. Note that model
configurations have not been added to config_grids, nor have
corresponding mapping or domain files been created. This is
because we do not yet know what the desired resolutions are for
the other components when combining with EHC at r025.